+++ app.js
@@ -5,7 +5,7 @@ var pkg = require("./package.json");
 var subscribeEndpoint = require("./prelaunch/subscribe-endpoint.js");
 var listSubscribersEndpoint = require("./prelaunch/list-subscribers-endpoint.js");
 var staticViewEndpoint = require("./staticViewEndpoint.js");
-var PostgresStatus = require("./PostgresStatus.js");
+var postgresStatusEndpoint = require("./status/postgres.js");
 var path = require("path");
 
 var app = express();
@@ -30,7 +30,7 @@ app.post("/subscribe", subscribeEndpoint);
 app.get("/du-er-paa-listen", staticViewEndpoint("prelaunch/subscription-succesful-view.html"));
 app.get("/list-subscribers", listSubscribersEndpoint);
 app.get("/om", staticViewEndpoint("about/view.html"));
-app.get("/pg-status", new PostgresStatus());
+app.get("/pg-status", postgresStatusEndpoint());
 
 app.use(function endpointNotFound(req, res) {
     res.fail(404);
